[action lists] - implemented versioning system for action lists - implemented better hashing / file management for cache

+++ src/main/java/io/github/chrisruffalo/pintle/resolution/ListController.java
@@ -137,7 +137,7 @@ public class ListController {
             // todo: log
             return;
         }
-        int loaded = processorOptional.get().process(processSourceEvent.getStoredListId(), processSourceEvent.getConfig(), source);
+        long loaded = processorOptional.get().process(processSourceEvent.getStoredListId(), processSourceEvent.getConfig(), source);
         logger.debugf("loaded %d items", loaded);
     }
